+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
   causing an error when the scheduler attempts to run jobs
   ([issue](https://github.com/hashicorp/boundary/issues/1309),
   [PR](https://github.com/hashicorp/boundary/pull/1310)).
+* Docker: update entrypoint script to handle more Boundary subcommands for 
+  better UX
 
 ## 0.3.0 (2021/06/08)